file organization
organized files to lib folder
